// src/core/base/common.types.ts
export type HTMLInputTypes =
    | 'button'
    | 'checkbox'
    | 'color'
    | 'date'
    | 'datetime-local'
    | 'email'
    | 'file'
    | 'hidden'
    | 'image'
    | 'month'
    | 'number'
    | 'password'
    | 'radio'
    | 'range'
    | 'reset'
    | 'search'
    | 'submit'
    | 'tel'
    | 'text'
    | 'textarea'
    | 'time'
    | 'url'
    | 'week'

export type FieldInputStateType =
    | 'dirty'
    | 'valid'
    | 'pristine'
    | 'focus'
    | 'open'
    | 'errors'
    /* even if required is not necessary an input flag and while it's a dupplicate value whitch is available in validation 
    it's convinient to have it as an input flag. All the flags are calculated as a separate object which can be useed outside the component and the field itself */
    | 'required'
    /* this will reset the style flags */
    | 'clear'